+++ cmd/generate.go
@@ -18,6 +18,7 @@ var (
 	structName  string
 	get         string
 	update      string
+	join        string
 	create      bool
 	test        bool
 )
@@ -35,7 +36,9 @@ func init() {
 	// TODO: add flag for table name
 	generateCMD.Flags().StringVarP(&packageName, "package", "p", "", "Name of repository package. default is 'repository'")
 	generateCMD.Flags().StringVarP(&get, "get", "g", "", "Get variables for GET functions in repository. ex: -g [ (var1,var2), (var2,var4), var3 ]")
-	generateCMD.Flags().StringVarP(&update, "update", "u", "", "Get variables for UPDATE functions in repository.  ex: -g [ [(byPar1,byPar2,...), (field1, field2)], ... ]")
+	generateCMD.Flags().StringVarP(&update, "update", "u", "", "Get variables for UPDATE functions in repository.  ex: -u [ [(byPar1,byPar2,...), (field1, field2)], ... ]")
+	generateCMD.Flags().StringVarP(&join, "join", "j", "", "Get variables for JOIN functions in repository.  ex: -j "+
+		"[ [(source_path, struct_name, variable_name_in_first_struct, join_on, join_type), (source_path, struct_name, variable_name_in_first_struct, join_on, join_type)], ... ]")
 	generateCMD.Flags().StringVarP(&structName, "struct-name", "n", "", "find struct with struct name in source file")
 	generateCMD.Flags().BoolVarP(&create, "create", "c", false, "Set to create CREATE function in repository")
 	generateCMD.Flags().BoolVarP(&test, "test", "t", false, "generate automatically tests for created repository")
@@ -52,7 +55,7 @@ func generate(_ *cobra.Command, _ []string) {
 	source = strings.Replace(source, " ", "", -1)
 	destination = strings.Replace(destination, " ", "", -1)
 
-	if get == "" && update == "" && !create {
+	if get == "" && update == "" && !create && join == "" {
 		log.Fatal("you must set at least one flag for get, update or create")
 	}
 
@@ -80,7 +83,19 @@ func generate(_ *cobra.Command, _ []string) {
 		updateVars = parser.ExtractUpdateVariables(update)
 	}
 
-	if err := repository.Generate(source, destination, packageName, structName, getVars, updateVars, create, test); err != nil {
+	var joinVars *[]structure.JoinVariables
+	if join != "" {
+		for strings.Contains(join, " ") {
+			join = strings.Replace(join, " ", "", -1)
+		}
+		join = strings.Replace(join, " ", "", -1)
+		if err := parser.ValidateJoinFlag(join); err != nil {
+			log.Fatal(err)
+		}
+		joinVars = parser.ExtractJoinVariables(join)
+	}
+
+	if err := repository.Generate(source, destination, packageName, structName, getVars, updateVars, joinVars, create, test); err != nil {
 		log.Fatal(err)
 	}
 }